+++ Migrations/Code/ClassAliasMap.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'Tx_Oelib_Domain_Model_GermanZipCode' => \OliverKlee\Oelib\Domain\Model\GermanZipCode::class,
+    'Tx_Oelib_Domain_Repository_GermanZipCodeRepository' => \OliverKlee\Oelib\Domain\Repository\GermanZipCodeRepository::class,
+];